Delete functionality added

+++ 04-rest-crud/02-spring-boot-rest-crud-employee/src/main/java/com/namancodes/crud_employee/rest/EmployeeRestController.java
@@ -45,8 +45,13 @@ public class EmployeeRestController {
 
     // delete employee
     @DeleteMapping("/employees/{employeeId}")
-    public void delete(@PathVariable int employeeId){
+    public String delete(@PathVariable int employeeId){
+        Employee tempEmployee = employeeService.findById(employeeId);
+        if(tempEmployee==null){
+            throw new RuntimeException("Employee id not found - "+employeeId);
+        }
         employeeService.deleteById(employeeId);
+        return "Deleted employee id - "+employeeId;
     }
 
     // create employee
